apps/api/requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.10.1

# Firebase and Google Cloud
firebase-admin==6.4.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-auth==2.40.3

# Database and ORM (if needed for complex queries)
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0

# Authentication and Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18

# AI and ML
openai==1.99.9
langchain==0.3.27
anthropic==0.8.1
textblob==0.19.0
tiktoken>=0.5.0
tenacity>=8.0.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0

# Blockchain and Crypto
web3==6.12.0
eth-account==0.10.0
cryptography==44.0.1

# HTTP Client and Async
httpx==0.28.1
aiohttp==3.12.15
requests==2.32.4

# Data Processing
pandas==2.1.4
numpy==1.25.2
python-dateutil==2.8.2

# Email and Communications
sendgrid==6.12.4
twilio==9.7.0

# Image and QR Code Processing
qrcode[pil]==8.2
pillow==10.3.0

# Payment Processing (Adyen)
Adyen==13.4.0
requests-oauthlib==2.0.0

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==2.35.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==24.3.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Environment and Configuration
python-dotenv==1.0.0
click==8.1.7

# CORS and Middleware
fastapi-cors==0.0.6

# Background Tasks
celery==5.3.4
redis==5.0.1

# Rate Limiting
slowapi==0.1.9

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.8 